Adds more dbt model file mockup functionality and continues refining structure.
