+++ airflow/dags/dev_experiments/rollout_standardized_dbt_model_dev.py
@@ -14,8 +14,14 @@ from sqlalchemy.engine.base import Engine
 
 from cc_utils.db import execute_result_returning_query, get_pg_engine
 from cc_utils.socrata import SocrataTable
-from cc_utils.file_factory import get_table_sqlalchemy_col_objects, write_lines_to_file
-from sources.tables import CHICAGO_CTA_TRAIN_STATIONS as SOCRATA_TABLE
+from cc_utils.file_factory import (
+    get_table_sqlalchemy_col_objects,
+    write_lines_to_file,
+    col_type_cast_formatter,
+    format_dbt_stub_for_intermediate_standardized_stage,
+    format_dbt_stub_for_intermediate_clean_stage,
+)
+from sources.tables import COOK_COUNTY_NEIGHBORHOOD_BOUNDARIES as SOCRATA_TABLE
 
 task_logger = logging.getLogger("airflow.task")
 
@@ -98,88 +104,18 @@ def get_columns_w_most_unique_vals(
     return cols_with_most_unique_vals
 
 
-def col_type_cast_formatter(col_name: str, sqlalch_col_type) -> str:
-    #     sqlalch_col_type: sqlalchemy.sql.sqltypes.*
-    if str(sqlalch_col_type).upper() == "BIGINT":
-        return f"        {col_name}::bigint AS {col_name},"
-    elif str(sqlalch_col_type).upper() == "INTEGER":
-        return f"        {col_name}::int AS {col_name},"
-    elif str(sqlalch_col_type).upper() == "SMALLINT":
-        return f"        {col_name}::smallint AS {col_name},"
-    elif str(sqlalch_col_type).upper() == "BOOLEAN":
-        return f"        {col_name}::boolean AS {col_name},"
-    elif str(sqlalch_col_type).upper() == "TEXT":
-        return f"        upper({col_name}::text) AS {col_name},"
-    elif str(sqlalch_col_type).upper().startswith("VARCHAR"):
-        return f"        upper({col_name}::varchar) AS {col_name},"
-    elif str(sqlalch_col_type).upper().startswith("CHAR"):
-        return f"        upper({col_name}::char) AS {col_name},"
-    elif str(sqlalch_col_type).upper().startswith("GEOMETRY"):
-        return f"        {col_name}::{str(sqlalch_col_type).upper()} AS {col_name},"
-    elif str(sqlalch_col_type).upper().startswith("DOUBLE_PRECISION"):
-        return f"        {col_name}::double precision AS {col_name},"
-    elif str(sqlalch_col_type).upper() == "DATE":
-        return f"        {col_name}::date AS {col_name},"
-    else:
-        return f"        {col_name}::MANUALLY_REPLACE (was {str(sqlalch_col_type)} AS {col_name},"
-
-
-def format_dbt_stub_for_standardized_stage(
-    table_name: str, engine: Engine, task_logger: Logger
-) -> List[str]:
-    table_cols = get_table_sqlalchemy_col_objects(
-        table_name=table_name, schema_name="data_raw", engine=engine
-    )
-    table_col_details = [{"name": col.name, "type": col.type} for col in table_cols]
-    file_lines = [
-        "{{ config(materialized='view') }}",
-        "{% set ck_cols = [",
-        "        REPLACE_WITH_COMPOSITE_KEY_COLUMNS",
-        "] %}",
-        "",
-        "WITH records_with_basic_cleaning AS (",
-        "    SELECT",
-        "        {{ dbt_utils.generate_surrogate_key(ck_cols) }} AS REPLACE_WITH_BETTER_id,",
-    ]
-
-    col_lines = []
-    for table_col_deet in table_col_details:
-        col_type = table_col_deet["type"]
-        col_name = table_col_deet["name"]
-        if col_name == "source_data_updated":
-            col_lines.append(f"        {col_name}::timestamptz AS {col_name},")
-        elif col_name == "ingestion_check_time":
-            col_lines.append(f"        {col_name}::timestamptz AS {col_name}")
-        else:
-            col_lines.append(col_type_cast_formatter(col_name=col_name, sqlalch_col_type=col_type))
-    file_lines.extend(col_lines)
-    file_lines.extend(
-        [
-            f"""    FROM {{{{ ref('{table_name}') }}}}""",
-            """    ORDER BY {% for ck in ck_cols %}{{ ck }}{{ "," if not loop.last }}{% endfor %}""",
-            ")",
-            "",
-            "",
-            "SELECT *",
-            "FROM records_with_basic_cleaning",
-            "ORDER BY {% for ck in ck_cols %}{{ ck }},{% endfor %} source_data_updated",
-        ]
-    )
-    task_logger.info(f"file_lines for table {table_name}")
-    for file_line in file_lines:
-        task_logger.info(f"    {file_line}")
-    return file_lines
-
-
-@task(retries=1)
+@task
 def make_dbt_standardized_model(table_name: str, conn_id: str, task_logger: Logger) -> None:
 
     engine = get_pg_engine(conn_id=conn_id)
-    std_file_lines = format_dbt_stub_for_standardized_stage(
-        table_name=f"temp_{table_name}", engine=engine, task_logger=task_logger
+    std_file_lines = format_dbt_stub_for_intermediate_standardized_stage(
+        table_name=table_name, engine=engine
     )
     file_path = Path(f"/opt/airflow/dbt/models/intermediate/{table_name}_standardized.sql")
     write_lines_to_file(file_lines=std_file_lines, file_path=file_path)
+    task_logger.info(f"file_lines for table {table_name}")
+    for file_line in std_file_lines:
+        task_logger.info(f"    {file_line}")
 
     task_logger.info(f"Leaving make_dbt_standardized_model")
