+++ airflow/dbt/models/intermediate/schema.yml
@@ -188,3 +188,11 @@ models:
   - name: cook_county_parcel_value_assessments_clean
     config:
       schema: clean
+
+  - name: chicago_cta_train_stations_standardized
+    config:
+      schema: clean
+
+  - name: cook_county_neighborhood_boundaries_standardized
+    config:
+      schema: clean
